Add get product endpoint
This endpoint allows us to query the database for a particular product's information.
This allows us to get the information from the database directly without having to get the information from the frontend when we try to edit the information of the product. This could prevent the user from editing the frontend informations and messing up the requests.